+++ CHANGES.rst
@@ -4,5 +4,6 @@
 * Better reporting of differences between the old and updated bibtex entries
 * More options and improved documentation 
 * Optimised mrlookup searches and bibtex field extraction.
+* Supported general format for values in bibtex files.
 
 **v1.0** July 2012 -- initial version.